initial card setup
