+++ src/App.js
@@ -10,23 +10,24 @@ import Signup from './pages/Signup';
 import { Routes, Route,  } from 'react-router-dom';
 import Contact from './pages/Contact';
 import Profile from './pages/ProfilePage';
+import IdeaBin from './pages/IdeaBin'
 
 
 function App() {
 
   return (
-    <ApolloProvider client={client}>
-    <Navbar />
-    <Routes>
-      <Route path="/" element={<HomePage />} />
-      <Route path="/login" element={<Login />} />
-      <Route path="/signup" element={<Signup />} />
-      <Route path="/profile" element={<Profile />} />
-      <Route path="/about" element={<About />} />
-      <Route path="/contact" element={<Contact />} />
-    </Routes>
-    </ApolloProvider>
-
+    // <ApolloProvider client={client}>
+    // <Navbar />
+    // <Routes>
+    //   <Route path="/" element={<HomePage />} />
+    //   <Route path="/login" element={<Login />} />
+    //   <Route path="/signup" element={<Signup />} />
+    //   <Route path="/profile" element={<Profile />} />
+    //   <Route path="/about" element={<About />} />
+    //   <Route path="/contact" element={<Contact />} />
+    // </Routes>
+    // </ApolloProvider>
+    <IdeaBin />
 
   );
 }
